services/autofact-scraper/requirements.txt
fastapi
uvicorn[standard]
playwright
beautifulsoup4
python-dotenv
